requirements.txt
fastapi
uvicorn[standard]
python-multipart
openai-whisper
openai>=1.0.0
gTTS
langdetect
python-dotenv
